# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Projeto Alergia",
    "version": "1.0.0",
    "description": "In order to successfully use the documentation, first create a new doctor and log in. By doing that you will be able to test the doctor's endpoints, register a new patient and later on, log in and test the patient's endpoints as well."
  },
  "servers": [
    {
      "url": "https://app-vacina-backend-production.up.railway.app/"
    }
  ],
...
  "paths": {
    "/doctor": {
      "post": {
        "tags": [
          "doctor"
        ],
        "description": "Add a doctor.",
        "operationId": "createDoctor",
        "requestBody": {
          "description": "Doctor´s information.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Doctor"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Email already in use."
          }
        }
      },
      "get": {
        "tags": [
          "doctor"
        ],
        "description": "List all doctors.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "operationId": "getDoctors",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DoctorResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "List not found."
          }
        }
      }
    },
    "/doctor/{id}": {
      "put": {
        "tags": [
          "doctor"
        ],
        "description": "Update doctor´s information.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "operationId": "updateDoctor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Doctor´s ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "New information data.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Doctor"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Doctor not found."
          }
        }
      },
      "get": {
        "tags": [
          "doctor"
        ],
        "description": "Find a doctor by ID.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "operationId": "getDoctorById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Doctor´s ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Doctor not found."
          }
        }
      }
    },
...
    "/doctor/patient-register": {
      "post": {
        "tags": [
          "doctor"
        ],
        "description": "Register a patient.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "operationId": "createPatient",
        "requestBody": {
          "description": "Patient´s information.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientRegistration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientRegistrationResponse"
                }
              }
            }
          },
          "409": {
            "description": "Email already in use."
          }
        }
      }
    },
...
    "/patient/{id}": {
      "patch": {
        "tags": [
          "patient"
        ],
        "description": "Update patient´s information.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "operationId": "updatePatient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Patient´s ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "New information data.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientResponse"
                }
              }
            }
          },
          "404": {
            "description": "Patient not found."
          }
        }
      },
      "get": {
        "tags": [
          "patient"
        ],
        "description": "Find a patient by ID.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "operationId": "getPatientById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Patient´s ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientResponse"
                }
              }
            }
          },
          "404": {
            "description": "Patient not found."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Doctor": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Mason B.",
            "minLength": 3
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "masonb@email.com"
          },
          "password": {
            "type": "string",
            "example": "1Ab@cdef"
          },
          "about": {
            "type": "string",
            "example": "Mason B., MD, PhD, is a board-certified neurosurgeon specializing in complex intracranial surgery including brain, pituitary, and skull base tumors, and brain aneurysms."
          },
          "CRM": {
            "type": "string",
            "example": "123456/SP"
          },
          "specialty": {
            "type": "string",
            "example": "Neurosurgery"
          },
          "photo": {
            "type": "string",
            "format": "url",
            "example": "url path"
          }
        }
      },
      "DoctorResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "uuid",
            "readOnly": true
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "example": "doctor",
            "readOnly": true
          },
          "about": {
            "type": "string"
          },
          "CRM": {
            "type": "string"
          },
          "specialty": {
            "type": "string"
          },
          "photo": {
            "type": "string",
            "format": "url"
          }
        }
      },
...
      "PatientRegistration": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Beatrice L.",
            "minLength": 3
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "beatrice.l@email.com"
          },
          "birthDate": {
            "type": "string",
            "format": "date",
            "example": "05/05/2005"
          },
          "dosage": {
            "type": "string",
            "enum": [
              "1:10",
              "1:100",
              "1:1.000",
              "1:10.000"
            ]
          },
          "allergies": {
            "type": "string",
            "enum": [
              "food",
              "wasp",
              "mite",
              "cat",
              "ball",
              "insect",
              "bee",
              "ant",
              "dog",
              "cockroach",
              "fungus",
              "candidiasis"
            ]
          },
          "frequency": {
            "type": "string",
            "enum": [
              "7 days",
              "2 weeks",
              "3 weeks",
              "4 weeks"
            ]
          },
          "method": {
            "type": "string",
            "enum": [
              "vaccine",
              "sublingual"
            ]
          },
          "startTreatment": {
            "type": "string",
            "format": "date",
            "example": "05/05/2023"
          },
          "endTreatment": {
            "type": "string",
            "format": "date",
            "example": "05/05/2025"
          }
        }
      },
      "PatientRegistrationResponse": {
        "type": "object",
        "properties": {
          "patient": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "readOnly": true
              },
              "name": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "birthDate": {
                "type": "string"
              },
              "doctorId": {
                "type": "string",
                "readOnly": true
              }
            }
          },
          "treatment": {
            "$ref": "#/components/schemas/Treatment"
          }
        }
      },
      "PatientUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Mason B."
          },
          "email": {
            "type": "string",
            "example": "masonb@email.com"
          },
          "password": {
            "type": "string",
            "example": "450Eht@6"
          },
          "photo": {
            "type": "string",
            "format": "url",
            "example": "url path"
          },
          "telephone": {
            "type": "string",
            "example": "(19)99999-9999"
          },
          "birthDate": {
            "type": "string",
            "format": "date",
            "example": "05/05/2005"
          }
        }
      },
      "PatientResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "photo": {
            "type": "string",
            "format": "url"
          },
          "telephone": {
            "type": "string",
            "example": "(19)99999-9999"
          },
          "birthDate": {
            "type": "string",
            "example": "05/05/2005"
          }
        }
      },
      "Treatment": {
        "type": "object",
        "properties": {
          "dosage": {
            "type": "string",
            "enum": [
              "1:10",
              "1:100",
              "1:1.000",
              "1:10.000"
            ]
          },
          "allergies": {
            "type": "string",
            "enum": [
              "food",
              "wasp",
              "mite",
              "cat",
              "ball",
              "insect",
              "bee",
              "ant",
              "dog",
              "cockroach",
              "fungus",
              "candidiasis"
            ]
          },
          "frequency": {
            "type": "string",
            "enum": [
              "7 days",
              "2 weeks",
              "3 weeks",
              "4 weeks"
            ]
          },
          "method": {
            "type": "string",
            "enum": [
              "vaccine",
              "sublingual"
            ]
          },
          "startTreatment": {
            "type": "string",
            "format": "date"
          },
          "endTreatment": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    }
  }
}